Extract _restart.py from _legacy.py
