+++ modules/helpers/__init__.py
@@ -32,4 +32,5 @@ from ._overlays import *  # noqa: F403
 from ._artifacts import *  # noqa: F403
 from ._plex_cache import *  # noqa: F403
 from ._misc import *  # noqa: F403
+from ._restart import *  # noqa: F403
 from ._paths import *  # noqa: F403
